+++ elastic4s-reactivestreams-pekko/src/test/scala/com/sksamuel/elastic4s/pekko/reactivestreams/ScrollPublisherVerificationTest.scala
@@ -1,4 +1,4 @@
-package com.sksamuel.elastic4s.pekko.http.streams
+package com.sksamuel.elastic4s.pekko.reactivestreams
 
 import org.apache.pekko.actor.ActorSystem
 import com.sksamuel.elastic4s.jackson.ElasticJackson
